requirements.txt
# Runtime dependencies for the single-file alconna-based maimaidx plugin
nonebot2[fastapi]>=2.3.0
nonebot-adapter-onebot>=2.4.6
nonebot-plugin-alconna>=0.59.4
nonebot-plugin-orm[sqlite]>=0.8.2
aiohttp>=3.13.1
async-lru>=2.0.5
maimai-py>=1.3.8
matplotlib>=3.9.4
numpy>=1.26,<2.4
pillow>=11.3.0
playwright>=1.41.0
pydantic<3
psutil>=6.1.0
python-dotenv>=1.0.0
typing-extensions>=4.12.0
uvicorn>=0.30.0
